Chore/prefer ts js extensions (#16)
* API Updates

* Update Connection String
