+++ apps/api/.env.schema
@@ -2,5 +2,7 @@
 PORT=3000
 
 # DB
-POSTGRES_URL=postgres://postgres:mypassword@localhost:5432/
+PGUSER=postgres
+PGPASSWORD=mypassword
+PGHOST=localhost:5432/
 PGDATABASE=postgres
