+++ apps/api/package.json
@@ -3,10 +3,10 @@
 	"private": true,
 	"version": "0.0.0",
 	"type": "module",
-	"main": "src/main.mts",
+	"main": "src/main.ts",
 	"scripts": {
-		"db-test": "npx tsx --env-file=.env ./src/db/index.mts",
-		"dev": "npx tsx watch --env-file=.env ./src/main.mts",
+		"db-test": "npx tsx --env-file=.env ./src/db/index.ts",
+		"dev": "npx tsx watch --env-file=.env ./src/main.ts",
 		"test": "npx tsx --test --test-reporter=spec --experimental-test-coverage ./src/tests/*"
 	},
 	"devDependencies": {
